Update to the ACDS 21.1std.0 tools era
Release testing environment:

Version 21.1.0 Build 842 10/21/2021 SJ Standard Edition

Release description:

This is the first update to the repo since the initial release.  There
are no substantial technical changes in this release, none of the IP
cores require modificiations or repairs.  There are a number of
cosmetic clean-up changes and a license change.

This release is being done to prepare the repo to migrate to the
Quartus Pro environment to support new components that have
dependencies on the Pro environment.  Moving forward the master branch
of this repo is going to focus specifically on the Quartus Pro
requirements and any future releases for the Quartus Standard
environment will likely branch from this release.

Specific changes in this release:

- Changed the project license from "MIT" to "MIT No Attribution" and
replaced all license text with the SPDX license identifier "MIT-0" in
all source files.

- Tested all development scripts and IP simulations with the 21.1std
version of the Intel FPGA tools.

- Added the release era "ACDS-STD-21-1-0-b842" to the repo maintenance
scripts.

- Cleaned up trailing white space line endings in a number of files.

- Updated README documentation to explain name change from Altera Qsys
tool to Intel Platform Designer tool and applied mild updates to the
tools references where the context was applicable.  Most references to
Qsys were left intact where they were incorporated into other names and
IP contexts.

- Updated references to GitHub organization change from "01org" to
"Intel" and updated URLs that can be used to obtain the repo.

- Adjusted the "add_instance" lines in the TCL scripts contained in
documentation/qsys_templates to remove the specific IP version to allow
the scripts to execute in later tools releases without modification.
This change allows the tools to instantiate the latest version of an IP
core instead of a specific version of the IP core.